Update error messages to include originating layer (#118)
* Add all error prefixes

* Rename ControllerIT to BaseControllerIT

* Move all testCreate methods to BaseServiceTest

+++ Backend/src/main/java/ca/mcgill/cooperator/service/CourseOfferingService.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class CourseOfferingService {
+public class CourseOfferingService extends BaseService {
     @Autowired CourseOfferingRepository courseOfferingRepository;
     @Autowired CourseRepository courseRepository;
 
@@ -38,7 +38,7 @@ public class CourseOfferingService {
             error.append("Course cannot be null!");
         }
         if (error.length() > 0) {
-            throw new IllegalArgumentException(error.toString().trim());
+            throw new IllegalArgumentException(ERROR_PREFIX + error.toString().trim());
         }
 
         CourseOffering co = new CourseOffering();
@@ -58,7 +58,7 @@ public class CourseOfferingService {
             error.append("Course Offering to update cannot be null!");
         }
         if (error.length() > 0) {
-            throw new IllegalArgumentException(error.toString().trim());
+            throw new IllegalArgumentException(ERROR_PREFIX + error.toString().trim());
         }
 
         if (year > 0) {
@@ -79,7 +79,7 @@ public class CourseOfferingService {
         CourseOffering c = courseOfferingRepository.findById(id).orElse(null);
         if (c == null) {
             throw new IllegalArgumentException(
-                    "Course Offering with ID " + id + " does not exist!");
+                    ERROR_PREFIX + "Course Offering with ID " + id + " does not exist!");
         }
         return c;
     }
@@ -89,7 +89,10 @@ public class CourseOfferingService {
         List<CourseOffering> co = courseOfferingRepository.findByCourse(c);
         if (co == null) {
             throw new IllegalArgumentException(
-                    "There are no course offerings for the course " + c.getName() + "!");
+                    ERROR_PREFIX
+                            + "There are no course offerings for the course "
+                            + c.getName()
+                            + "!");
         }
         return co;
     }
@@ -105,19 +108,25 @@ public class CourseOfferingService {
     public List<CourseOffering> getAllCourseOfferings() {
         List<CourseOffering> co = ServiceUtils.toList(courseOfferingRepository.findAll());
         if (co == null) {
-            throw new IllegalArgumentException("There are no course offerings!");
+            throw new IllegalArgumentException(ERROR_PREFIX + "There are no course offerings!");
         }
         return co;
     }
 
     @Transactional
     public CourseOffering deleteCourseOffering(CourseOffering co) {
+        if (co == null) {
+            throw new IllegalArgumentException(
+                    ERROR_PREFIX + "Course offering to delete cannot be null!");
+        }
         courseOfferingRepository.delete(co);
+
         Course c = co.getCourse();
         List<CourseOffering> cos = c.getCourseOfferings();
         cos.remove(co);
         c.setCourseOfferings(cos);
         courseRepository.save(c);
+
         return co;
     }
 }
